chore!: Move client into protocol and rename builders in adapters
